Add a converter for enum flags that can also be boolean.
A new converter class is added that makes it possible for enum flags to define a conversion from booleans to one of the enumeration values. This is in addition to the enumeration conversions.

Fields that define such a converter can be used with the --flag and --noflag
forms in addition to their --flag=value enumeration style.

--
MOS_MIGRATED_REVID=93972718

GitOrigin-RevId: 173ee825277d64c8a852f25b6d6e0a85917342ec
Change-Id: Icffd5b9a30a80121d78833715f7d257a3bde9416