Rationalize copyright headers
The headers were modified with
`find . -type f -exec 'sed' '-Ei' 's|Copyright 201([45]) Google|Copyright 201\1 The Bazel Authors|' '{}' ';'`
And manual edit for not Google owned copyright. Because of the nature of ijar, I did not modified the header of file owned by Alan Donovan.

The list of authors were extracted from the git log. It is missing older Google contributors that can be added on-demand.

--
MOS_MIGRATED_REVID=103938715

GitOrigin-RevId: f88f4d81acead9ef3e2f73d9a44afd76c2c239c1
Change-Id: I24bf6dcdbf1c77ea6ae5eaef65072ec9727d6fb7