Add a resource parsing action.
Part of 3 proposed new actions:
- parsing action
- merging action
- validating action

Dependencies (directData and transitiveData)
expect the symbol files. If the merge action
produces the symbol files, then each merge
action depends on each other. Instead, produce
it in an action with just source resources as
prereqs to allow more parallelism.

Technically, we don't need a manifest as part
of the parameters. I debated about whether
to introduce a basic version of
UnvalidatedAndroidData or not.

--
MOS_MIGRATED_REVID=128599714

GitOrigin-RevId: fcd684775f21d8fc38228368c6ac4b2a0213def5
Change-Id: Ib7a5ac5845e6be340ea825a82e5bd4d6069fa6c6