Add a flag to enable desugaring java.* classes by rewriting package names
java.* and sun.* classes cannot be desugared directly because there
are hard coded restrictions in the JVM that prevent using loading or
using reflection on them.  Add a --core_library flag that rewrites
package names to have a __desugar__ prefix when reading them, and
strips the prefix when writing them back out.

Test: builds
Change-Id: I706b6b09d8a100476aef5fd3fffe7153278e9ad5