Create "internal" category of command-line options.
This is intended to be used for "flags" which should never appear on the
command line - things like configuration distinguishers, which are used
internally and must be part of the build options, but should always be set
to their default value at the top level.

This is already a convention within Bazel, but doesn't actually work the way
Bazel expects - flags with spaces can be set by simply escaping or quoting
the spaces so that word splitting will not break on them. This means they
can also be matched by config_settings, which pass a single string.
Forbidding the parser from matching these flags solves both of these
unintended cases.

Existing cases like this have also been converted to internal.

--
PiperOrigin-RevId: 150497246
MOS_MIGRATED_REVID=150497246

GitOrigin-RevId: a5aa916dc8c0e99c8fc91893246381f72b7066a9
Change-Id: I782990abd7e12010fe8b9ef35475b676e89011f8