Expand Invocation FlagPolicies on expansion flags.
For SetValue and UseDefault policies on expansion flags or flags with implicitRequirements, expand the policy into policy for each of its sub-flags. For SetValue, this addresses the issue with policies on expansion flags with overridable=true not actually letting user flags overrride the expansion. For UseDefault, this formalizes the behavior where UseDefault will wipe all user-provided flags that expand from a banned expansion flag, and will allow later work to guarantee that a later policy can override the expansion policy's subflags.

Since expansion flags do not have value, break if the invocation policy uses AllowValue or DisallowValue on an expansion flag.

PiperOrigin-RevId: 151718539

GitOrigin-RevId: 706bafe7aa17ed6ef1187986af3ba749559fb232
Change-Id: I75e326362f23733f0cf1df4d12c0c0b107b6e813