Add new exception for wrapping parser construction failures
The exception is unchecked. The reasoning is that errors during parser construction should not occur, and when they do occur it is an internal error like a failed assertion.

This allows casual uses of the options parser to stay oblivious to the possibility of failures, consistent with how DuplicateOptionDeclarationException is currently [not] handled. At the same time, the dispatcher can catch the exception to fail gracefully (by printing to stdout instead of a log file) when parser construction fails for any reason.

RELNOTES: None

PiperOrigin-RevId: 151839620

GitOrigin-RevId: ba03c11bb8155bca74ddb4539acd0c0f2105da11
Change-Id: I3470a02ac6f54cfee521c78e1bd232dab2f38103