Add integration test for --all_incompatible_changes flag conflicts
This adds a new warning when the same flag is expanded to by multiple expansion flags. This extends an existing suite of warnings, e.g. for when an expansion flag conflicts with an explicit option.

The blaze canonicalize-flags command now takes a new flag --show_warnings. This flag causes any warnings encountered while parsing the given command line to be printed to stderr.

RELNOTES: blaze canonicalize-flags now takes a --show_warnings flag

PiperOrigin-RevId: 152186672

GitOrigin-RevId: e11775c2394fc48ac7fe5b632b47ae952dd552b4
Change-Id: Id14310d27000d5c7cbe6fe4ec236c3591b9ff1be