Fix an expansion flag issue where it expands to a repeatable flag.
There's no reason an expansion flag should not expand to multiple values for a repeatable flag (a flag with allowMultiple set to true.) If this expansion flag is set in a SetValue policy, group its repeatable subflags into a single SetValue per subflag.

For an overridable SetValue policy on an expansion, any repeatable flag that it expands to should append its value, and not override the user's original values.

PiperOrigin-RevId: 153233784
GitOrigin-RevId: ae1d0dec94e4f4b7066ce1c65d4f035907540586
Change-Id: Ia9bfedb601ba8d5c4484e6d590491d3712b4fb51