Make Desugar work for relative dumpDirectory paths.
Before CL 155913466, Desugar was not working when dumpDirectory was a relative path because of an extra/incorrect resolve() that ended up duplicating such a relative path.

CL 155913466 fixed this by converting Desugar's dumpDirectory to an absolute path. However, CL 156257767 introduced a check that Desugar's dumpDirectory must match the one in Java's InnerClassLambdaMetafactory, which broke the case of a relative path again.

Therefore, this CL undoes the conversion to absolute path again (so that Desugar's path matches that in InnerClassLambdaMetafactory). Instead, to make relative paths work, it drops the unnecessary/incorrect call to resolve().

A regression test is left for a future CL (I don't have a dev environment set up that would allow me to easily write one).

RELNOTES: None.
PiperOrigin-RevId: 156852882
GitOrigin-RevId: 6bb44ec649fe91dd1a3e33e02edfb3caf6d4b7e8
Change-Id: I4b83516a8ee1e70d5cb3817dbceeec9f6d75759e