Move InvocationPolicy to the options parser package.
It was originally included in runtime due to external dependencies, and
a desire to keep the options parser a general options library. These
dependencies have been or will be removed, and there are plenty of other
general flag libraries.

InvocationPolicy is fundamentally acting on the properties of this
specific OptionsParser and needs proper access to it for the proper
solution to a number of existing bugs, which means having access to
things that should be package private.

PiperOrigin-RevId: 158523111
GitOrigin-RevId: f39dc6fc707cca9496d5baef5549bb7c24933776
Change-Id: I9f71233464721a0ab4c885d8e70f1d4e6fbd4127