Automated rollback of commit 317e0de2113f0361c9dbf98653ceda4efea3f3e2.
*** Original change description ***

Access interface constants to explicitly trigger the execution of interface
initializers. The problem is that when we desugar default methods, the static
intializer of interface will not be executed. The JVM spec says that if an
interface has default methods, then when it is loaded, it will also be
initialized. After we desugar such an interface, its default methods are
removed, and when we load the interface, the <clinit> will not be executed.

This CL checks whether an interface has default me...

***

ROLLBACK_OF=159496992
NO_SQ=PURE_ROLLBACK

RELNOTES: None
PiperOrigin-RevId: 159545752
GitOrigin-RevId: 2d008e824e4c5c39f00f21cfa879eed61d5b500a
Change-Id: I0bd3917d683f058c0ebe8ba5c3bf7b10feb55722