Fold OptionUsageRestrictions into OptionDocumentationCategory and OptionMetadataTags.
These are similar, no need to have both fields. Removing the "DOCUMENTED" default, the absence of UNDOCUMENTED will be used instead.

Since requiring a documentation category for undocumented options doesn't make sense, list that as one of the OptionDocumentationCategories, but list HIDDEN and INTERNAL as part of OptionMetadata. These options should list UNDOCUMENTED as their category.

PiperOrigin-RevId: 161515674
GitOrigin-RevId: 456adb2267343a4e2e64f082e77169c18f9e6060
Change-Id: Ia7dcbd5d38d4caa2d6d60f25b5a1f0c23735ca22