Add the option filter enums in java, with equivalence check.
In order to break out the proto dependency of the OptionsParser, add a java-only enum that users of the options parser can depend on, instead of requiring the dependency on the protobuf. The protocol buffer definition is kept in order to be used for command line reporting in the BEP, so a test is added to make sure that the two are kept in sync.

Also add some clearer guidance for how to go about selecting, and adding, option categories and tags.

Migrating the current uses of the proto enum to this one will happen in a follow-up change.

PiperOrigin-RevId: 161971077
GitOrigin-RevId: d1c4d22f548880f1c6b159735040dbba4c513eeb
Change-Id: Iedc53b1746be2fa9af49fbbc1b884cf9d0f15dfd