Use cached values for option converter types.
This requires us to have OptionsData for all usage messages, since static functionality is being removed, but this should already have been the case. It was added as an optional argument when the expansion function feature was added, but there is actually no reason not to require it, as the public interface for usage text was already computing the optionsData anyway.

PiperOrigin-RevId: 165386893
GitOrigin-RevId: c69dbf8ed4893066d0474f87c0a8ddae00e4ee58
Change-Id: Iba12221dbcf759ab090c3c66509c89055e820fd1