Move option tag checks to compile time.
The rationality checks failed as RuntimeExceptions that weren't getting reported properly because the options parser is created too early in Bazel's lifetime to be reported properly to standard error. This was a minor annoyance for authors of new options, and there was nothing gained by waiting until runtime to check these values.

RELNOTES: None
PiperOrigin-RevId: 166395759
GitOrigin-RevId: 41f2974d50aaafe713d938866ee17b03c58bec00
Change-Id: I90393753c7cc58020a515dc510cfdf8046e4c0a3