Replace referrals to options by their name to option definitions.
Now that we have a standard way of referring to an option, remove all of the places that we were referring to them by their name. Since options can have multiple names, this is more clear and provides the additional information needed to understand the option. It also stops the habit of requesting unqualified strings, which was hard to read.

RELNOTES: None.
PiperOrigin-RevId: 168254584
GitOrigin-RevId: 1dce09721f8361240bbf056fd508f1ac5fdcfd32
Change-Id: I5ca50358d96ed613c544248920ca51cbe856842e