Pass the UnparsedOptionValues when setting or adding option values.
Stop breaking the value apart to be recombined later. Also stop using OptionValueDescriptions as though we have a final option when expanding flags for invocation policy. These values are explicitly the output from parsing the expansion strings, not yet converted or combined with other values of the same flag.

After this change, only UnparsedOptionValueDescription should be used when strings of flags are parsed, and OptionValueDescription should be used when the final version of a flag is created or used.

PiperOrigin-RevId: 168688063
GitOrigin-RevId: a8c0c8dfad38437bc617b4b66d271bc0beb1b7c1
Change-Id: I159b1a1a076704bc3b90760b8c8ad628a135830a