Throw away the confusing UnparsedOptionValueDescription name.
We get UnparsedValues after ... parsing the options. So that doesn't
make sense. What was meant was that it wasn't converted to the final
value.

In an effort to make this distinction more clear, this change will
make the terminology more consistent. The `--foo=bar` step is
"parsing" and the `bar -> Object` step is "converting" (it is, in
fact, done by Converters).

RELNOTES: None.
PiperOrigin-RevId: 168852847
GitOrigin-RevId: fb153cd83c6f801271048ec1d62d17a68562376d
Change-Id: I90065b60724e698b93e685c045a21f901adb3c14