Track the origin of an option in the option instance, not its final value.
A single instance of an option has a single origin, but the final value only has a single origin if it has a single value. For multi-valued options, it is wrong to expect that the final value of an option will have a single parent. Track the option parents (which option expanded to the current instance, if any) in the right place, with the ParsedOptionDescription.

Also fix some inconsistent spelling of 'dependent,' in favor of the American English standard.

RELNOTES: None.
PiperOrigin-RevId: 169487515
GitOrigin-RevId: 3a0df3cb0637d71dfcf0add7057332c09cd508c5
Change-Id: Ic187397f8204aea3638b109b71d590920e968bdb