Clean up InvocationPolicy's use of OptionDescription.
OptionDescription is basically a hack to get the expansion data for options from outside the options parser, but it was being used at various points of invocation policy enforcement. In order to correctly track option origin, we only want to get this information once. Do it during the invocation policy expansion stage, not at enforcement, so that we track the information of the option's origin in the original invocation policy passed to the enforcer, not the expanded one.

PiperOrigin-RevId: 171661669
GitOrigin-RevId: ca74482825e0c0ca5d119eceab74ba4292428557
Change-Id: Ib57ed5c67054c051a420507506807a8eb6809f47