Expand implicitRequirements in the location of the option that required it.
Removes the special casing of implicit requirements. Accumulating them and parsing them at the end of the parse() function was never enough to actually guarantee that the value not be replaced. I've gone through all options with implicit requirements to make sure that the expectation is checked after options parsing, so this change should be relatively safe.

Implicit requirements is still a broken concept - they don't actually expand based on the value given, so a user that is explicitly NOT setting a flag might unwittingly be setting all the requirements for that unset flag. Removing it fully requires redesigning or removing the flags that set it, though, so for now we are standardizing the behavior so that it behaves like any other expansion options, just one with a value.

Also consolidate the deprecated wrapper option behavior into the expansion work. It will soon be removed entirely, but for now it can get grouped in with the expansion logic, so that its differences are more explicit.

RELNOTES: None.
PiperOrigin-RevId: 171957502
GitOrigin-RevId: faf2db22277bd0b074a0a8c5cd4533a735eff31a
Change-Id: I84a1461624207d3d43786ebc69d17ab6f16f9ec8